dynamicitem now has a proper data structure
